backend/requirements.txt
fastapi
uvicorn
torch
torchvision
mlflow
albumentations
Pillow
prometheus-fastapi-instrumentator
python-multipart
numpy
scikit-image
python-dotenv